requirements.txt
# Python dependencies for repository scripts
PyYAML>=6.0